campus-genius-backend/chat_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-engineio==4.8.1
pymongo==4.6.1
python-dotenv==1.0.0
motor==3.3.2
python-multipart==0.0.6
pydantic==2.5.2 